requirements.txt
# Shared for comparison runner and web app
python-dotenv>=1.0.0
pydub>=0.25.0
numpy>=1.20.0
Flask>=3.0.0
pyopenssl>=23.0.0

# Google Cloud Speech-to-Text v2
google-cloud-speech>=2.20.0
google-auth>=2.0.0

# AWS Transcribe Streaming
amazon-transcribe>=0.6.0

# Azure Speech (REST)
aiohttp>=3.8.0
